server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "argon2-cffi>=25.1.0",
    "bitsandbytes>=0.49.2",
    "datetime>=6.0",
    "fastapi[standard]>=0.129.0",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]

[tool.basedpyright]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv"
]
extraPaths = ["."]

reportMissingImports = "error"
reportUnknownMemberType = "warning"
reportUntypedFunctionDecorator = "none"

venv = ".venv"
venvPath = "."
